renamed QuellCache to match class naming convention
